requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
netmiko==4.2.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.5
python-dotenv==1.0.0
python-telegram-bot==20.7
